Move common ComplexType logic to base class
