+++ ClosedXML.IO.CodeGen/Model/TopLevel/ComplexType.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using ClosedXML.IO.CodeGen.Model.Elements;
 
 namespace ClosedXML.IO.CodeGen.Model.TopLevel;
@@ -8,8 +9,6 @@ namespace ClosedXML.IO.CodeGen.Model.TopLevel;
 /// </summary>
 public abstract class ComplexType : IReferencable
 {
-    internal const string Prefix = "CT_";
-
     /// <summary>
     /// Name of the complex type.
     /// </summary>
@@ -23,5 +22,75 @@ public abstract class ComplexType : IReferencable
     /// </summary>
     public required bool? Mixed { get; init; }
 
+    internal void Generate(CodeBuilder code, string namespaceField)
+    {
+        code.StartMethod("void Parse{0}(string elementName)", Name);
+        code.OpenBrace();
+        foreach (var oneOfAttribute in Attributes)
+        {
+            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
+            {
+                attribute.Generate(code);
+            }
+            else
+            {
+                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
+            }
+        }
+
+        GenerateParseMethod(code, namespaceField);
+        CallListener(code);
+        code.CloseBrace();
+
+        AddPartialMethodSignature(code, Name);
+    }
+
     internal abstract void GenerateParseMethod(CodeBuilder code, string namespaceField);
+
+    private void CallListener(CodeBuilder code)
+    {
+        code.WriteIndent().Append("On").AppendComplexType(Name).Append("Parsed(");
+        var isFirst = true;
+        foreach (var oneOfAttribute in Attributes)
+        {
+            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
+            {
+                if (!isFirst)
+                    code.Append(", ");
+                code.AppendVariable(attribute.Name!);
+                isFirst = false;
+            }
+            else
+            {
+                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
+            }
+        }
+
+        code.Append(");").EndLine();
+    }
+
+    private void AddPartialMethodSignature(CodeBuilder code, string typeName)
+    {
+        code.EndLine();
+        code.WriteIndent().Append($"partial void On").AppendComplexType(typeName).Append("Parsed(");
+
+        var isFirst = true;
+        foreach (var oneOfAttribute in Attributes)
+        {
+            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
+            {
+                if (!isFirst)
+                    code.Append(", ");
+
+                code.AppendSimpleType(attribute).Append(" ").AppendVariable(attribute.Name!);
+                isFirst = false;
+            }
+            else
+            {
+                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
+            }
+        }
+
+        code.Append(");").EndLine();
+    }
 }
