+++ ClosedXML.IO.CodeGen/Model/TopLevel/ComplexTypeElement.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace ClosedXML.IO.CodeGen.Model.TopLevel;
+﻿namespace ClosedXML.IO.CodeGen.Model.TopLevel;
 
 /// <summary>
 /// <c><![CDATA[<xsd:complexType/>]]></c> inside <c><![CDATA[<xsd:schema/>]]></c>. It doesn't have
@@ -15,21 +13,6 @@ public class ComplexTypeElement : ComplexType, INode
 
     internal override void GenerateParseMethod(CodeBuilder code, string namespaceField)
     {
-        code.StartMethod("void Parse{0}(string elementName)", Name)
-            .OpenBrace();
-
-        foreach (var oneOfAttribute in Attributes)
-        {
-            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
-            {
-                attribute.Generate(code);
-            }
-            else
-            {
-                throw new NotImplementedException($"Attribute group '{attributeGroup}' read not implemented.");
-            }
-        }
         code.AddLine($"_reader.Close(elementName, {namespaceField});");
-        code.CloseBrace();
     }
 }
