+++ ClosedXML.IO.CodeGen/Model/TopLevel/ComplexTypeSequence.cs
@@ -29,20 +29,6 @@ public class ComplexTypeSequence : ComplexType, INode
 
     internal override void GenerateParseMethod(CodeBuilder code, string namespaceField)
     {
-        code.StartMethod("void Parse{0}(string elementName)", Name);
-        code.OpenBrace();
-        foreach (var oneOfAttribute in Attributes)
-        {
-            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
-            {
-                attribute.Generate(code);
-            }
-            else
-            {
-                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
-            }
-        }
-
         var min = Sequence.Occurrences.Min ?? 1;
         var max = Sequence.Occurrences.Max ?? 1;
         if (min == 1 && max == 1)
@@ -65,56 +51,5 @@ public class ComplexTypeSequence : ComplexType, INode
         }
 
         code.AddLine($"_reader.Close(elementName, {namespaceField});");
-        CallListener(code);
-        code.CloseBrace();
-
-        AddPartialMethodSignature(code, Name);
-    }
-
-    private void CallListener(CodeBuilder code)
-    {
-        code.WriteIndent().Append("On").AppendComplexType(Name).Append("Parsed(");
-        var isFirst = true;
-        foreach (var oneOfAttribute in Attributes)
-        {
-            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
-            {
-                if (!isFirst)
-                    code.Append(", ");
-                code.AppendVariable(attribute.Name!);
-                isFirst = false;
-            }
-            else
-            {
-                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
-            }
-        }
-
-        code.Append(");").EndLine();
-    }
-
-    private void AddPartialMethodSignature(CodeBuilder code, string typeName)
-    {
-        code.EndLine();
-        code.WriteIndent().Append($"partial void On").AppendComplexType(typeName).Append("Parsed(");
-
-        var isFirst = true;
-        foreach (var oneOfAttribute in Attributes)
-        {
-            if (oneOfAttribute.TryPickT1(out var attribute, out var attributeGroup))
-            {
-                if (!isFirst)
-                    code.Append(", ");
-
-                code.AppendSimpleType(attribute).Append(" ").AppendVariable(attribute.Name!);
-                isFirst = false;
-            }
-            else
-            {
-                throw new NotImplementedException($"Attribute group ({attributeGroup.RefName}) not yet implemented.");
-            }
-        }
-
-        code.Append(");").EndLine();
     }
 }
